Allows both old and new StarGAN models to be used...
The old (2022) StarGAN model (capacity 1, 2-input only) requires the
domain the be already channelized and concatenated with the source
image. But the recent StarGAN models allow passing 2 or 3 inputs
and the source/target domain labels can be in the shape of [b, d],
which is easier, and they do not have to be concatenated with the
source image.